refactor(worker): extract download/network helpers into downloads.rs (sc-1636)
Move the HTTP download machinery out of lib.rs into downloads.rs: the
SnapshotFile/HuggingFaceSnapshot/DownloadContext/DownloadProgress types,
the HF snapshot resolve+download (download_snapshot), the URL download
path (download_source_url + lora/model variants, content-length, DNS
validation), and the progress reporting (report_download_progress,
download_progress_payload).

Same pattern: `use super::*;`, glob re-import at the root. pub(crate)
widened to what crosses the module boundary — the model_jobs handlers
build a DownloadContext and call download_snapshot/download_source_url,
and tests inspect HuggingFaceSnapshot/SnapshotFile/DownloadProgress, so
those structs' fields and HuggingFaceSnapshot::{resolve,total_bytes} /
DownloadProgress::new are pub(crate). tests.rs imports repointed to
super::downloads. Behavior-preserving. lib.rs 1,829 -> 1,326 lines.
